pyproject.toml
[project]
name = "kprl"
version = "0.1.0"
description = "RealLive SEEN.TXT and image format tools (Python port of kprl/vaconv/rlxml)"
readme = "README.md"
authors = [
    { name = "Nightdavisao", email = "nightdavisao27@protonmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0",
]

[project.scripts]
kprl = "kprl.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["kprl"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
